Move StringWrapperClass to leoAPI.py. This is a breaking change
